src/requirements.txt
termcolor
mysql-connector-python
PySocks
scrypt